Add files via upload
Made 2 path finding codes, one with dynamic environment and one with static. In dynamic the obstacles are initially at same positions as static. But these obstacles move forward. In this version all obstacle move with same speed. I will make it more dynamic in that sense in future. However no i can compare between a static and dynamic obstacle places. Make a dataset of the date on various cases and then feed the data to ML algorithms, so that I can get a model for a particular environment. This code is a basic version only and just describes the idea I have to counter dynamic environment obstacles.
06-05-2022
10:54 AM