Add files via upload
This was my project for AI Course that I was undergoing in my 4th Sem of BE Computer Engineering. I had an idea of my project, so I made a solo team rather than a duo. Completed the project well before time and also grabbed a research paper related to my topic from Dr Singara Singh Kasana. The project was well documented and presented. In future I will be working on the same topic under guidance of Dr Singara sir.
Submitted and presented on 20-05-2022 at 4:00 PM